+++ googleapis_auth/lib/googleapis_auth.dart
@@ -18,14 +18,15 @@
 /// {@canonicalFor auth_functions.refreshCredentials}
 /// {@canonicalFor client_id.ClientId}
 /// {@canonicalFor exceptions.AccessDeniedException}
+/// {@canonicalFor exceptions.ServerRequestFailedException}
 /// {@canonicalFor exceptions.RefreshFailedException}
 /// {@canonicalFor exceptions.UserConsentException}
 /// {@canonicalFor response_type.ResponseType}
 /// {@canonicalFor service_account_credentials.ServiceAccountCredentials}
 library googleapis_auth;
 
-export 'src/access_credentials.dart';
-export 'src/access_token.dart';
+export 'src/access_credentials.dart' show AccessCredentials;
+export 'src/access_token.dart' show AccessToken;
 export 'src/auth_client.dart';
 export 'src/auth_functions.dart';
 export 'src/client_id.dart';